Move some interfaces to mm-core
