+++ mindmaps-graql/src/main/java/io/mindmaps/graql/internal/pattern/VarImpl.java
@@ -22,10 +22,7 @@ import com.google.common.collect.Maps;
 import io.mindmaps.concept.ResourceType;
 import io.mindmaps.graql.ValuePredicate;
 import io.mindmaps.graql.Var;
-import io.mindmaps.graql.admin.Conjunction;
-import io.mindmaps.graql.admin.Disjunction;
-import io.mindmaps.graql.admin.ValuePredicateAdmin;
-import io.mindmaps.graql.admin.VarAdmin;
+import io.mindmaps.graql.admin.*;
 import io.mindmaps.graql.internal.gremlin.MultiTraversal;
 import io.mindmaps.graql.internal.gremlin.VarTraversals;
 import io.mindmaps.graql.internal.pattern.property.*;
